// SIC/SIC Admin/src/types/component.d.ts
export type DropdownOptionType = {
	name: string | number;
	key: string | number;
};